+++ YogaSMCPane/YogaSMCPane.swift
@@ -16,98 +16,6 @@ let DYTCCommand = ["L", "M", "H"]
 let thinkLEDCommand = [0, 0x80, 0xA0, 0xC0]
 let thinkBatteryName = ["BAT_ANY", "BAT_PRIMARY", "BAT_SECONDARY"]
 
-// from https://ffried.codes/2018/01/20/the-internals-of-the-macos-hud/
-@objc enum OSDImage: CLongLong {
-    case kBrightness = 1
-    case brightness2 = 2
-    case kVolume = 3
-    case kMute = 4
-    case volume5 = 5
-    case kEject = 6
-    case brightness7 = 7
-    case brightness8 = 8
-    case kAirportRange = 9
-    case wireless2Forbid = 10
-    case kBright = 11
-    case kBrightOff = 12
-    case kBright13 = 13
-    case kBrightOff14 = 14
-    case Ajar = 15
-    case mute16 = 16
-    case volume17 = 17
-    case empty18 = 18
-    case kRemoteLinkedGeneric = 19
-    case kRemoteSleepGeneric = 20 // will put into sleep
-    case muteForbid = 21
-    case volumeForbid = 22
-    case volume23 = 23
-    case empty24 = 24
-    case kBright25 = 25
-    case kBrightOff26 = 26
-    case backlightonForbid = 27
-    case backlightoffForbid = 28
-    /* and more cases from 1 to 28 (except 18 and 24) */
-}
-
-func getBoolean(_ key: String, _ io_service: io_service_t) -> Bool {
-    guard let rvalue = IORegistryEntryCreateCFProperty(io_service, key as CFString, kCFAllocatorDefault, 0) else {
-        return false
-    }
-
-    return rvalue.takeRetainedValue() as! Bool
-}
-
-func getNumber(_ key: String, _ io_service: io_service_t) -> Int {
-    guard let rvalue = IORegistryEntryCreateCFProperty(io_service, key as CFString, kCFAllocatorDefault, 0) else {
-        return -1
-    }
-
-    return rvalue.takeRetainedValue() as! Int
-}
-
-func getString(_ key: String, _ io_service: io_service_t) -> String? {
-    guard let rvalue = IORegistryEntryCreateCFProperty(io_service, key as CFString, kCFAllocatorDefault, 0) else {
-        return nil
-    }
-
-    return rvalue.takeRetainedValue() as! NSString as String
-}
-
-func getDictionary(_ key: String, _ io_service: io_service_t) -> NSDictionary? {
-    guard let rvalue = IORegistryEntryCreateCFProperty(io_service, key as CFString, kCFAllocatorDefault, 0) else {
-        return nil
-    }
-
-    return rvalue.takeRetainedValue() as? NSDictionary
-}
-
-func sendBoolean(_ key: String, _ value: Bool, _ io_service: io_service_t) -> Bool {
-    return (kIOReturnSuccess == IORegistryEntrySetCFProperty(io_service, key as CFString, value as CFBoolean))
-}
-
-func sendNumber(_ key: String, _ value: Int, _ io_service: io_service_t) -> Bool {
-    return (kIOReturnSuccess == IORegistryEntrySetCFProperty(io_service, key as CFString, value as CFNumber))
-}
-
-func sendString(_ key: String, _ value: String, _ io_service: io_service_t) -> Bool {
-    return (kIOReturnSuccess == IORegistryEntrySetCFProperty(io_service, key as CFString, value as CFString))
-}
-
-// from https://github.com/alin23/Lunar/blob/master/Lunar/Data/Hotkeys.swift
-func showOSD(_ prompt: String, _ img: NSString? = nil, _ duration: UInt32 = 1000, _ priority: UInt32 = 0x1f4) {
-    guard let manager = OSDManager.sharedManager() as? OSDManager else {
-        os_log("OSDManager unavailable", type: .error)
-        return
-    }
-
-    manager.showImage(
-        atPath: img ?? "/System/Library/CoreServices/OSDUIHelper.app/Contents/Resources/kBright.pdf",
-        onDisplayID: CGMainDisplayID(),
-        priority: priority,
-        msecUntilFade: duration,
-        withText: prompt as NSString)
-}
-
 class YogaSMCPane : NSPreferencePane {
     let io_service = IOServiceGetMatchingService(kIOMasterPortDefault, IOServiceMatching("YogaVPC"))
     let defaults = UserDefaults(suiteName: "org.zhen.YogaSMC")!
