+++ YogaSMCUtils/YogaSMC-Bridging-Header.h
@@ -1,17 +1,17 @@
 //
-//  YogaSMCNC-Bridging-Header.h
-//  YogaSMCNC
+//  YogaSMC-Bridging-Header.h
+//  YogaSMC
 //
 //  Created by Zhen on 10/14/20.
 //  Copyright © 2020 Zhen. All rights reserved.
 //
 
-#ifndef YogaSMCNC_Bridging_Header_h
-#define YogaSMCNC_Bridging_Header_h
+#ifndef YogaSMC_Bridging_Header_h
+#define YogaSMC_Bridging_Header_h
 
 #import <Foundation/Foundation.h>
 #import <OSD/OSDManager.h>
 
 #import "../YogaSMC/YogaSMCUserClient.h"
 
-#endif /* YogaSMCNC_Bridging_Header_h */
+#endif /* YogaSMC_Bridging_Header_h */
